Merger the .h and .cc of all plugins in RecoParlicleFlow/PFProducer
